+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add Dark Wizard, Mega Spider and Dark Spider mobs
  - Add Webbing spell (for spider mobs)
  - Add /mmob clear command, for getting rid of magic mobs
+ - Add "Magic.use.<wand>" permission list, for controlling individual wand permissions.
 
 ## 6.0.11